PUR-70 feat: add OpinionList component for Home
