+++ src/components/domain/Opinion/OpinionList/OpinionList.style.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 40px;
+  margin-top: 24px;
+`;
