+++ src/components/domain/Opinion/OpinionList/index.ts
@@ -1,0 +1,3 @@
+import OpinionList from './OpinionList';
+
+export default OpinionList;
