+++ src/components/template/OpinionListTemplate/OpinionListTemplate.tsx
@@ -2,48 +2,12 @@ import { useRouter } from 'next/router';
 
 import Button from '@/components/common/Button';
 import Text from '@/components/common/Text';
-import OpinionItem from '@/components/domain/Opinion/OpinionItem';
-import type { OpinionType } from '@/constants/types';
-import useIntersectionObserver from '@/hooks/useIntersectionObserver';
-import { useGetOpinionList } from '@/query-hooks/opinion';
+import OpinionList from '@/components/domain/Opinion/OpinionList';
 
 import * as styles from './OpinionListTemplate.style';
 
 const OpinionListTemplate = () => {
   const router = useRouter();
-
-  const {
-    data: opinionDataList,
-    fetchNextPage: fetchNextOpinionList,
-    hasNextPage,
-  } = useGetOpinionList({
-    size: 10,
-    status: 'IN_PROGRESS',
-  });
-
-  const handleFetchNextOpinion = (entries: IntersectionObserverEntry[]) => {
-    if (hasNextPage && entries[0].isIntersecting) fetchNextOpinionList();
-  };
-
-  const { targetRef } = useIntersectionObserver({
-    onIntersect: handleFetchNextOpinion,
-  });
-
-  const opinionList =
-    opinionDataList?.pages
-      .reduce<OpinionType[]>(
-        (previous, { content: currentOpinions }) => [
-          ...previous,
-          ...currentOpinions,
-        ],
-        [],
-      )
-      .flat() || [];
-  <styles.BottomWrap>
-    <Button sizeType="large" isFilled isRound>
-      컨텐츠 제안하기
-    </Button>
-  </styles.BottomWrap>;
 
   return (
     <>
@@ -54,22 +18,7 @@ const OpinionListTemplate = () => {
         <Text fontStyleName="body2R">
           투표 기간 끝나서 완료되었던 것들 여기에 몰아넣을거에요
         </Text>
-        <styles.ListWrap
-          ref={(element) => {
-            if (element?.lastElementChild)
-              targetRef.current = element.lastElementChild;
-          }}
-        >
-          {opinionList.map((opinion) => (
-            <OpinionItem
-              key={opinion.id}
-              id={opinion.id}
-              title={opinion.title}
-              thumbnailUrl={opinion.thumbnailUrl}
-              endDate={opinion.endDate}
-            />
-          ))}
-        </styles.ListWrap>
+        <OpinionList status="FINISHED" />
       </styles.Wrapper>
       <styles.BottomWrap>
         <Button
